+++ src/claude/lib/usage/constants.ts
@@ -1,39 +1,16 @@
-import type { LimitKind } from "@genesiscz/utils/ai/providers/account-features";
-
-export const BUCKET_LABELS: Record<string, string> = {
-    five_hour: "Session (5h)",
-    seven_day: "Weekly (all)",
-    seven_day_opus: "Weekly (Opus)",
-    seven_day_sonnet: "Weekly (Sonnet)",
-    seven_day_oauth_apps: "Weekly (OAuth)",
-    extra_usage: "Extra usage",
-};
-
-export const BUCKET_PERIODS_MS: Record<string, number> = {
-    five_hour: 5 * 60 * 60 * 1000,
-    seven_day: 7 * 24 * 60 * 60 * 1000,
-    seven_day_opus: 7 * 24 * 60 * 60 * 1000,
-    seven_day_sonnet: 7 * 24 * 60 * 60 * 1000,
-    seven_day_oauth_apps: 7 * 24 * 60 * 60 * 1000,
-};
-
 /**
- * `LimitWindow.kind` for a claude bucket (orchestrator amendment 3). The scoped weekly
- * windows keep their model in `scopeModel`; only the legacy projection still speaks
- * `weekly_all` / `five_hour`.
+ * The bucket vocabulary itself lives in the anthropic-sub plugin (`buckets.ts`), because
+ * the usage mapper that produces `LimitWindow[]` is inside `src/utils/**` and may not
+ * import `@app/*`. This file keeps the claude TUI's own presentation: Ink colours, the
+ * percent thresholds and the reset-imminence rule.
  */
-export const BUCKET_KIND_MAP: Record<string, LimitKind> = {
-    five_hour: "session",
-    seven_day: "weekly",
-    seven_day_opus: "scoped",
-    seven_day_sonnet: "scoped",
-    seven_day_oauth_apps: "weekly",
-    extra_usage: "credit",
-};
-
-export function bucketKind(bucket: string): LimitKind {
-    return BUCKET_KIND_MAP[bucket] ?? "weekly";
-}
+export {
+    BUCKET_KIND_MAP,
+    BUCKET_LABELS,
+    BUCKET_PERIODS_MS,
+    bucketKind,
+    VISIBLE_BUCKETS,
+} from "@genesiscz/utils/ai/providers/plugins/anthropic-sub/buckets";
 
 export const BUCKET_THRESHOLD_MAP: Record<string, "session" | "weekly"> = {
     five_hour: "session",
@@ -42,8 +19,6 @@ export const BUCKET_THRESHOLD_MAP: Record<string, "session" | "weekly"> = {
     seven_day_sonnet: "weekly",
     seven_day_oauth_apps: "weekly",
 };
-
-export const VISIBLE_BUCKETS = ["five_hour", "seven_day", "seven_day_opus", "seven_day_sonnet", "seven_day_oauth_apps"];
 
 export const BUCKET_COLORS: Record<string, string> = {
     five_hour: "\x1b[36m", // cyan
